Move email-assigned-users logic into serializer
